Handle reseting the method promotion better when a accessor is set too and from an array
The 'from array' case isn't really all that nice. Its not a supported use case just yet. If an array accessor is called as if to set with a number, that is interpreted as a call to get the value.
I think I should support a more explicit syntax to cover for this case. Something like `obj.property.set(1)`, which will set no matter what.